wip: use vite lib instead
