+++ applications/tari_indexer/web_ui/src/main.tsx
@@ -22,7 +22,7 @@
 
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "@tari-project/ootle-web-ui-theming/dist/theme/theme.css";
+import "@tari-project/ootle-web-ui-theming/theme.css";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import App from "./App";
 import ErrorPage from "./routes/ErrorPage";
