+++ applications/theming/lib/index.ts
@@ -1,0 +1,9 @@
+//   Copyright 2026 The Tari Project
+//   SPDX-License-Identifier: BSD-3-Clause
+/// <reference types="vite/client" />
+import "./theme.css";
+import { tariBg, tariPurple, grey, teal, gothic, blue, orange, green, red } from "./theme/colors";
+import { componentSettings, light, dark } from "./theme/tokens";
+export * from "./theme/augmentations";
+
+export { componentSettings, light, dark, tariBg, tariPurple, grey, teal, gothic, blue, orange, green, red };
