--- applications/theming/src/index.ts
@@ -1,6 +1,0 @@
-//   Copyright 2026 The Tari Project
-//   SPDX-License-Identifier: BSD-3-Clause
-
-export * from "./theme/augmentations";
-export * from "./theme/colors";
-export * from "./theme/tokens";
