--- applications/theming/src/theme/colors.ts
@@ -1,117 +1,0 @@
-//   Copyright 2026 The Tari Project
-//   SPDX-License-Identifier: BSD-3-Clause
-export const tariBg = {
-    50: "#fafaff",
-    100: "#f3f3fc",
-    200: "#c2c1dd",
-    300: "#a7a7c1",
-    400: "#8d8da7",
-    500: "#74748d",
-    600: "#5c5b74",
-    700: "#45445b",
-    800: "#2f2e44",
-    900: "#161625",
-    950: "#0a0a13",
-};
-export const tariPurple = {
-    "50": "#F6F0FE",
-    "100": "#DECBFC",
-    "200": "#C7A7FA",
-    "300": "#9996f6",
-    "400": "#645fec",
-    "500": "#813BF5",
-    "600": "#6830CB",
-    "700": "#4F26A1",
-    "800": "#361B77",
-    "900": "#1D114D",
-    "950": "#040723",
-};
-export const grey = {
-    50: "#f7f7f8",
-    100: "#eeeef0",
-    200: "#d8d9df",
-    300: "#b6b7c3",
-    400: "#8f91a1",
-    500: "#717286",
-    600: "#5b5d6e",
-    700: "#4a4a5a",
-    800: "#28282E",
-    900: "#161617",
-    950: "#090719",
-};
-export const teal = {
-    50: "#F1FAF9",
-    100: "#DBF2F1",
-    200: "#BBE6E5",
-    300: "#88D3D3",
-    400: "#55B9BB",
-    500: "#3A9DA0",
-    600: "#338187",
-    700: "#2F686F",
-    800: "#2D575D",
-    900: "#294A50",
-    950: "#172F35",
-};
-export const gothic = {
-    50: "#F3F8F8",
-    100: "#E0EBED",
-    200: "#C5D8DC",
-    300: "#9CBCC4",
-    400: "#63929E",
-    500: "#517D89",
-    600: "#466874",
-    700: "#3D5661",
-    800: "#384A52",
-    900: "#324047",
-    950: "#1E292E",
-};
-export const blue = {
-    50: "#EFF7FF",
-    100: "#DAEDFF",
-    200: "#BEE0FF",
-    300: "#91CEFF",
-    400: "#5DB2FD",
-    500: "#318EFA",
-    600: "#2172EF",
-    700: "#195CDC",
-    800: "#1B4BB2",
-    900: "#1C428C",
-};
-export const orange = {
-    50: "#FDF7EF",
-    100: "#FBEBD9",
-    200: "#F5D4B3",
-    300: "#ECA86A",
-    400: "#E88F4F",
-    500: "#E2712D",
-    600: "#D35923",
-    700: "#AF441F",
-    800: "#8C3720",
-    900: "#712F1D",
-};
-export const green = {
-    50: "#EEFBF3",
-    100: "#D6F5E1",
-    200: "#B0EAC7",
-    300: "#7CD9A7",
-    400: "#47C083",
-    500: "#229B62",
-    600: "#168552",
-    700: "#126A45",
-    800: "#115438",
-    900: "#0F452F",
-    950: "#07271B",
-};
-export const red = {
-    50: "#FEF3F2",
-    100: "#FEE4E2",
-    200: "#FECECA",
-    300: "#FCACA5",
-    400: "#F87C71",
-    500: "#EF5244",
-    600: "#DC3526",
-    700: "#B9291C",
-    800: "#A2281D",
-    900: "#7F251D",
-    950: "#450F0A",
-};
