--- applications/theming/tsconfig.prod.json
@@ -1,4 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {}
-}
